Rename module and add doctest tests
